# apps/web/app/admin/recovery-handoffs/crm/route.ts
import { NextResponse } from 'next/server';
import { adminApiFetch } from '@/lib/adminApiServer';

function redirectTarget(request: Request) {
  const url = new URL(request.headers.get('referer') ?? '/admin/dashboard', request.url);
  url.pathname = '/admin/dashboard';
  return url;
}

export async function POST(request: Request) {
  const target = redirectTarget(request);
  const res = await adminApiFetch('/v1/admin/stats/recovery-handoffs/crm', {
    method: 'POST',
    headers: { 'content-type': 'application/json' },
    body: JSON.stringify({ force: true }),
    cache: 'no-store',
  });
  const body = (await res.json().catch(() => null)) as {
    ok?: boolean;
    syncedCount?: number;
    skipped?: string | null;
    error?: string;
    detail?: string;
  } | null;

  if (!res.ok) {
    target.searchParams.set(
      'recoveryCrmError',
      body?.error ?? body?.detail ?? 'recovery_handoff_crm_unavailable',
    );
    return NextResponse.redirect(target, { status: 303 });
  }

  target.searchParams.set(
    'recoveryCrm',
    body?.skipped === 'no_crm_handoffs'
      ? 'no_crm_handoffs'
      : body?.skipped === 'already_synced'
        ? 'already_synced'
        : body?.skipped === 'no_due_crm_handoffs'
          ? 'no_due_crm_handoffs'
          : String(body?.syncedCount ?? 0),
  );
  return NextResponse.redirect(target, { status: 303 });
}