# packages/contracts/openapi/startup-graveyard.v1.yaml
openapi: 3.1.0
info:
  title: Startup Graveyard API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: sg_access
      description: Host-only HttpOnly access cookie; public production uses the configured __Host- cookie name.
  schemas:
...
    PlatformSnapshot:
      type: object
      required:
        [
          createdAt,
          triggerType,
          mockMode,
          queuedCount,
          oldestQueuedAgeMinutes,
          runningCount,
          staleRunningCount,
          failedCount,
          completedLastHour,
          alertCount,
          criticalAlertCount,
          warningAlertCount,
          infoAlertCount,
          workerStatus,
          workerConsecutiveErrors,
          workerLastProcessedAt,
        ]
      properties:
        createdAt: { type: string }
        triggerType:
          type: string
          enum: [manual, scheduled]
        mockMode: { type: boolean }
        queuedCount: { type: integer, minimum: 0 }
        oldestQueuedAgeMinutes: { type: integer, minimum: 0, nullable: true }
        runningCount: { type: integer, minimum: 0 }
        staleRunningCount: { type: integer, minimum: 0 }
        failedCount: { type: integer, minimum: 0 }
        completedLastHour: { type: integer, minimum: 0 }
        alertCount: { type: integer, minimum: 0 }
        criticalAlertCount: { type: integer, minimum: 0 }
        warningAlertCount: { type: integer, minimum: 0 }
        infoAlertCount: { type: integer, minimum: 0 }
        workerStatus:
          type: string
          enum: [disabled, idle, processing, error, stopped]
        workerConsecutiveErrors: { type: integer, minimum: 0 }
        workerLastProcessedAt: { type: string, nullable: true }
...
paths:
...
  /v1/admin/stats/platform-snapshot:
    post:
      summary: Capture the current platform diagnostics as a point-in-time snapshot in the admin audit log
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      responses:
        '200':
          description: Snapshot captured
          content:
            application/json:
              schema:
                type: object
                required: [ok, auditId, snapshot]
                properties:
                  ok:
                    type: boolean
                  auditId:
                    type: string
                    format: uuid
                  snapshot:
                    $ref: '#/components/schemas/PlatformSnapshot'
        '401':
          description: Unauthorized
        '503':
          description: Backing service unavailable
...
  /v1/admin/stats/recovery-member-email:
    post:
      summary: Deliver due fallback member recovery notification emails to the configured SMTP mailer
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                retryIntervalHours:
                  type: integer
                  minimum: 0
                  maximum: 336
                force:
                  type: boolean
      responses:
        '200':
          description: Member recovery emails delivered or skipped when nothing due remains
        '401':
          description: Unauthorized
        '502':
          description: Member recovery email delivery failed
        '503':
          description: Recovery member email transport is not configured or admin API disabled
  /v1/admin/stats/recovery-playbook:
    post:
      summary: Run the Team Workspace recovery playbook to refresh outreach and fan out owner/member email, CRM sync, webhook, and Slack escalation
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                retryIntervalHours:
                  type: integer
                  minimum: 0
                  maximum: 336
                force:
                  type: boolean
      responses:
        '200':
          description: Recovery playbook executed and returned per-channel summary
        '401':
          description: Unauthorized
        '502':
          description: Recovery playbook finished with at least one failed outbound channel
        '503':
          description: Backing service unavailable
...
  /v1/admin/stats/recovery-handoffs/crm:
    post:
      summary: Deliver due handed-off CRM recovery outreach rows to the configured CRM API and persist the external case id
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                retryIntervalHours:
                  type: integer
                  minimum: 0
                  maximum: 336
                force:
                  type: boolean
      responses:
        '200':
          description: CRM handoffs synced or skipped when nothing due remains
        '401':
          description: Unauthorized
        '502':
          description: CRM sync failed
        '503':
          description: Recovery CRM API is not configured or admin API disabled
...
  /v1/admin/stats/recovery-handoffs/slack:
    post:
      summary: Deliver dead-letter handed-off admin recovery outreach rows to the configured Ops Slack webhook
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
      responses:
        '200':
          description: Dead-letter handoffs delivered to Ops Slack or skipped when nothing needs alerting
        '401':
          description: Unauthorized
        '502':
          description: Slack delivery failed
        '503':
          description: Recovery Slack webhook is not configured or admin API disabled
...